+++ packages/terra-list/src/ListSubsection.jsx
@@ -90,6 +90,9 @@ const ListSubsection = ({
   const isListItemDropped = useRef();
   const draggedItemindex = useRef();
 
+  /* eslint-disable-next-line no-param-reassign */
+  delete customProps?.isTabFocusDisabled;
+
   useEffect(() => {
     if (!isCollapsible || !isCollapsed) {
       if (Array.isArray(children)) {