+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Focus is not maintained on Dropped list item when using Drag and Drop.
+
 ## 4.67.0 - (November 13, 2023)
 
 * Changed