+++ packages/terra-list/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Unreleased
- * Updated 
-   * Added `ariaDescription` prop in list
+ * Changed 
+   * Reverted `Removed ariaDescription prop from list`.
 
 * Fixed
   * After drag and drop in `terra-modal-manager` the dropped list item does not receive focus.