+++ packages/terra-list/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+
  * Changed 
    * Reverted `Removed ariaDescription prop from list`.