+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.68.0 - (December 11, 2023)
+
  * Changed 
    * Reverted `Removed ariaDescription prop from list`.